Move compat into utils/extension
