+++ app/src/main/kotlin/com/absinthe/libchecker/utils/FreezeUtils.java
@@ -4,7 +4,7 @@ import android.content.pm.ApplicationInfo;
 
 import androidx.annotation.Nullable;
 
-import com.absinthe.libchecker.compat.VersionCompat;
+import com.absinthe.libchecker.utils.extensions.VersionCompat;
 
 import java.lang.reflect.Field;
