+++ app/src/main/kotlin/com/absinthe/libchecker/viewmodel/DetailViewModel.kt
@@ -14,18 +14,18 @@ import com.absinthe.libchecker.annotation.*
 import com.absinthe.libchecker.api.ApiManager
 import com.absinthe.libchecker.api.bean.LibDetailBean
 import com.absinthe.libchecker.api.request.LibDetailRequest
+import com.absinthe.libchecker.bean.LibChip
 import com.absinthe.libchecker.bean.LibStringItemChip
 import com.absinthe.libchecker.bean.StatefulComponent
-import com.absinthe.libchecker.compat.VersionCompat
 import com.absinthe.libchecker.constant.AbilityType
 import com.absinthe.libchecker.constant.GlobalValues
-import com.absinthe.libchecker.bean.LibChip
 import com.absinthe.libchecker.constant.librarymap.IconResMap
 import com.absinthe.libchecker.database.Repositories
 import com.absinthe.libchecker.ui.fragment.detail.LocatedCount
 import com.absinthe.libchecker.ui.fragment.detail.MODE_SORT_BY_SIZE
 import com.absinthe.libchecker.utils.LCAppUtils
 import com.absinthe.libchecker.utils.PackageUtils
+import com.absinthe.libchecker.utils.extensions.VersionCompat
 import com.absinthe.libchecker.utils.harmony.ApplicationDelegate
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
